Delete confirmed

+++ src/app/components/pedidos/pedidos.component.ts
@@ -35,10 +35,26 @@ export class PedidosComponent {
     });
   }
   deletePedido(pedido: Pedido): void {
-    this._pedidosService.delete('/pedido', pedido).subscribe((data) => {
-      this._pedidosService.get('/pedido').subscribe((pedidos) => {
-        this.pedidos = pedidos.pedidos;
-      });
+    Swal.fire({
+      title: 'Estas Segura Margarita?',
+      text: `Estas segura de borrar el pedido ${pedido.pedidoId} de ${pedido.nombre}`,
+      showConfirmButton: true,
+      showCancelButton:true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Confirmar',
+      cancelButtonText: 'Cancelar',
+      background:'#aeaeae',
+    }
+    ).then(resp=>{
+      if(resp.isConfirmed)
+      {
+        this._pedidosService.delete('/pedido', pedido).subscribe((data) => {
+          this._pedidosService.get('/pedido').subscribe((pedidos) => {
+            this.pedidos = pedidos.pedidos;
+          });
+        });
+      }
     });
   }
   buscarPorFecha(date: string): void {
